requirements.txt
sqlalchemy
flask
pymysql
flask_sqlalchemy
PyJWT
flask_cors
python-dotenv
gunicorn